requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.1
pandas==1.2.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
